app/requirements.txt
Django==1.9.7
djangorestframework==3.3.3
psycopg2==2.6.1
pudb==2016.2
django-filter==0.13.0
Markdown==2.6.6
requests==2.10.0
django-anymail==0.4.2
uWSGI==2.0.13.1
